Fix: silence -Wsign-compare in libopendkim/tests buffer-fill loops
t-signperf and t-verifyperf reuse the same local 'c' (int) for both
getopt() (whose -1 sentinel requires int) and the body-buffer fill
loop (which compares against sizeof body, size_t).  Retyping c would
break the getopt usage, so cast to size_t at the two comparison
sites in the fill loop and leave a note.

t-test113 declares its own c only for one buffer-fill loop, so just
declare it as size_t.

+++ libopendkim/tests/t-verifyperf.c
@@ -300,9 +300,12 @@ main(int argc, char **argv)
 	srandom(time(NULL));
 
 	/* prepare a random body buffer */
-	for (c = 0, w = 0; c < sizeof body; c++)
+	/* c is reused as getopt()'s int (with -1 sentinel) elsewhere in
+	 * this function, so cast at the size_t comparisons rather than
+	 * retyping the local. */
+	for (c = 0, w = 0; (size_t) c < sizeof body; c++)
 	{
-		if (w >= 75 && c < sizeof body - 2)
+		if (w >= 75 && (size_t) c < sizeof body - 2)
 		{
 			body[c++] = '\r';
 			body[c++] = '\n';
